chore: add v10 upgrade handler (#1043)
* chore: add v10 upgrade handler

* rm v7

* Update app/app.go


* rm test file
